webapi/tct-extra-html5-tests/dom/documents/the-document-object/document_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu.Xin <xinx.liu@intel.com>

-->

<meta charset="utf-8">
<title>document test: attributes basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/dom.html#documents">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var element = document.createElement("div");

  test(function() {
    assert_true("execCommand" in document, "execCommand method exists");
  }, "Test checks that document.execCommand method exists");

  test(function() {
    assert_true("queryCommandEnabled" in document, "queryCommandEnabled method exists");
  }, "Test checks that document.queryCommandEnabled method exists");

  test(function() {
    assert_true("queryCommandIndeterm" in document, "queryCommandIndeterm method exists");
  }, "Test checks that document.queryCommandIndeterm method exists");

  test(function() {
    assert_true("queryCommandState" in document, "queryCommandState method exists");
  }, "Test checks that document.queryCommandState method exists");

  test(function() {
    assert_true("queryCommandSupported" in document, "queryCommandSupported method exists");
  }, "Test checks that document.queryCommandSupported method exists");

  test(function() {
    assert_true("queryCommandValue" in document, "queryCommandValue method exists");
  }, "Test checks that document.queryCommandValue method exists");

  test(function() {
    assert_true("isContentEditable" in element, "isContentEditable attribute exists");
  }, "Test checks that HTML element isContentEditable attribute exists");
</script>
